backend/requirements.txt
numpy
astropy
django
psycopg2
gunicorn
whitenoise
requests
django-cors-headers
git+https://github.com/ivan-katkov/django-custom-query.git#egg=django-custom-query
git+https://github.com/jazzband/django-silk.git
django-filter
djangorestframework
django-debug-toolbar
serpy
drf_ujson2
drf-yasg
django-rest-marshmallow
drf-flex-fields
pglast
sqlparse